Split out ScraperExecutor
